requirements.txt
langchain>=0.2.0
langgraph>=0.2.0
python-dotenv>=1.0.1
pydantic>=2.7.0
flask>=3.0.0
langchain-community>=0.3.0
langchain-openai>=0.1.0
dashscope>=1.24.0
tavily-python>=0.5.0
pymilvus>=2.4.0
protobuf>=5.27.2,<7
